+++ src/Spinner.tsx
@@ -1,8 +1,23 @@
 import React, { FC } from 'react';
 
-export const Spinner: FC = () => {
+export interface SpinnerProps {
+  large?: boolean;
+  small?: boolean;
+}
+
+export const Spinner: FC<SpinnerProps> = ({ large, small }) => {
+  let smallSpinner = true;
+
+  if (large !== undefined) smallSpinner = large;
+  if (small !== undefined) smallSpinner = small;
+
   return (
-    <div className="spinner large spinner-border" role="status">
+    <div
+      className={
+        smallSpinner ? 'spinner spinner-border' : 'spinner large spinner-border'
+      }
+      role="status"
+    >
       <span className="visually-hidden">Loading...</span>
     </div>
   );